Moved AppConfig type to gateway.service
